MAINT: moved argparser to new module
